Fix slash usage while calling wc_gzd_sanitize_html_text_field. Prevent db updates from running/queuing right after install due to wrong db_version check.

+++ includes/updates/woocommerce-gzd-update-1.6.3.php
@@ -4,6 +4,10 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
+if ( ! function_exists( 'WC_trusted_shops' ) ) {
+	return;
+}
+
 // Reset Trusted Shops Code Options
 update_option( 'woocommerce_gzd_trusted_shops_integration_mode', 'standard' );
 update_option( 'woocommerce_gzd_trusted_shops_product_sticker_code', WC_trusted_shops()->trusted_shops->get_product_sticker_code( false ) );
